Project Scaffolding
Begin project scaffolding, using Vite with react-ts template for the front-end and a basic express application for the backend. Dependencies have been added but are not comprehensive. Goal is to get docker-compose up and running.

[Test]
Connecting to frontend/backend is not quite working, but `docker-compose up` runs.  Docker works but probably need to tweak the commands.