Setup Authentication Service on Backend.
Start the foundations of an authentication service for use throughout the application. This will allow us to register, login, logout, and change password.

Added bcrypt password hashing to the User model on a pre-hook save to simplify the process. Also created a few helpers -- createUser, findByEmail, etc.

Needed to migrate from inlined and inferred types in mongoose to declared types. It wasn't picking up comparePassword on instances due to differences in `this` within the static object vs. creating an instance from scratch. tl;dr -- with a few type assertions everything is working now.

[Problem]
Need to be able to create users, hash passwords.

[Test]
Unit tests are passing.