Implement local _and_ JWT strategies for authentication
We're going to use JWT in a cookie. After some debate, we'll include just the user's ID in the JWT and lookup user with each API call. This will involve a caching layer (not yet implemented) that will reduce db hits.

Reworked authentication middleware -- the login middleware was unnecessary. Instead we'll use auth middleware to check and attach the user. Providing two options, required and optional auth.