Test coverage for authentication routes
Refactored index so that we can create a common test app to be used in-app and in-test. This requires a bit of code to identify whether the app is running or if the module is being imported to prevent us from trying to kick off the real app.

Created a test helper that manages app creation and shoudl hopefully simplify API requests.

[Problem]
Need basic test coverage for authentication

[Test]
Unit tests passing.