Implement auth context/provider/hooks
Using two contexts here --- one for auth data and a separate one for auth actions. This allows the `useAuth` hook to have a clean separation around manipulating auth state. Added several utilities, such as authStatus and nullable type. Some re-naming and organization to keep consistency across files as well.

Intention is for UI and other areas of frontend to consume useAuth only. We'll wrap the whole app in the AuthProvider component. Going to add unit tests in the next commit.