Implement basic auth and login forms
Hooked the application up to the useAuth context, implementing basic forms for registration and login. We have individual components set up for TextInput, but will need to keep working to build out Buttons, Tabs, etc.  Getting closer to the design comp though.

Tested in Docker and flow is working correctly.