Error handling and unit testing for auth forms.
Added basic error handling for registration and login when network/api failures occur.
Will expand this error handling in the future, but for now it at least gets the user informed.
We'll probably want to pull out error strings as well, into their own strings/constants file.

Added unit testing for new components. It's not comprehensive, but it gets the coverage up.
Tested the individual components as well as the registration/login forms.