Handle CSRF Token in authClient.
We want to make request details 'invisible' to the rest of the application. In addition, the auth context is really for UI... so move CSRF token out of auth context and into the apiClient module itself. It should last the session, but we have a mechanism to clear it if need be.