Implement JWT refresh on front-end.
When a call fails with a 401, attempt a JWT token refresh and then try again. If the call fails again let it ride.

Had to re-work apiClient a little, modifying how fetch is invoked to cleanly abstract the refresh, but this simplified version should handle most of our use cases.