Implement RBAC (role-based access control)
Using definition of household roles, create (potential) permissions for each document and assign to different roles. This may be updated in the future. A service provides methods for checking whether the user has permission based on their role, and we make middleware to enforce the permission.

This is only hooked up to some routes on household for now. Had to update household and household tests to get everything working correctly.

Unit tests are passing.