Implement invitations on the backend.
Invitations are generated in the POST to household/invitation. They can be redeemed at invitations/redeem. New model was created for invitation to hold things. No cron job to automatically expire... but when redeeming we check the expiration anyway.

Prepped shared library with schema, types, and constants to share between front and backend.