Align Landing page with application.
The landing page now uses our components for several of the elements, providing consistency across the application. Added a "FooterLogo" for now... may adjust the way we handle the header/footer logo in the future.

StatCard was huge, so made some modifications to allow size variants, which seems to work great and align with comps. Needed a change to section container to allow for hero-width elements too.