Responsive authenticated session header.
Header includes <nav> for navigation, a desktop and mobile-ready profile menu, and a mobile <nav> as well.

Ran into an issue with brutal-shadow-dark as it uses text-primary as it's shadow color. So I added shadow-black and shadow-white (along with sm and md variations). Using this on the mobile menu for now.

Will be refactoring to clean up the MainHeader.